refactor: create separate file for button class
